More refactoring.
No more macros folder. Moved to Common.
Updated readme.
Fixed rebuilding of all state's dependencies every time it's compiled.
Misc changes.
