+++ README.md
@@ -1,29 +1,40 @@
 Worldsong
 =========
 
-Write code. Hit compile. Tab into the running application. See your changes updated, running at native speed. Smile as your state data isn't decimated in the process.
-
-This is the dream of Worldsong.
+Worldsong is a framework that enables safe, easy runtime editing of your application's behaviour, without sacrificing performance or state data integrity.
 
 
-How to use
+Quickstart
 ----------
 
 1. Navigate to tools/scripts/os_setup.
 2. Run the script for your OS. (Currently only gnu/linux)
-3. When it finishes, you can run _launch_ in the root directory.
-4. While it's running, edit processes/graphics/graphics.rs, and run _compile_ in the same directory.
-5. Refocus the window.
+3. When it finishes, you can run _launch_ in the worldsong root directory.
+4. While it's running, try editing processes/graphics/graphics.rs, and run _compile_ in the same directory.
+5. Refocus the Worldsong window.
 6. Smile!
 
-Each subsystem has its own custom _compile_. Have fun!
+Each subsystem has its own custom _compile_. 
+
+To add your own functionality:
+
+1. Run 'add --name yourstructname --editor youreditor' in the structs directory
+2. Add some data definitions.
+3. Run 'add --name yourprocessname --editor youreditor' in the processes directory
+4. Write some code that modifies yourstructname
+5. Add your process to the appropriate schedule, passing it yourstructname.
+
+Have fun!
 
 
 What it's made of
 -----------------
 
-#####common Library:
-Stores data and functionality common to other subsystems, including as state data and third-party libraries.
+#####State Structs
+Modules that describe state data.
+
+#####State Library
+A library automatically generated from state structs, and their requested dependencies.
 
 #####Kernel:
 Initializes the state data, hotloads the scheduler, and sends the state data to it.
@@ -32,53 +43,37 @@ Initializes the state data, hotloads the scheduler, and sends the state data to 
 The main loop. Sets the conditions and timing for when schedules should be run.
 
 #####Schedule(s):
-List which processes should be run, in order, when the schedule is called.
+List which processes should be run, and in what order, when the schedule is called.
 
-#####Process(es):
-Modify state data. These babies do the actual work! (Hint: they're just functions)
+#####Processes:
+Modify state data. These babies do the actual work!
 
+#####Common Library:
+Stores data and functionality common to other subsystems. Boring API stuff.
 
 *In practice, it works something like this:*
 
->![It's something like this, anyway.](http://i.imgur.com/PZJEnhB.png)
+>![It's something like this, anyway.](http://i.imgur.com/Rac2pZq.png)
 
 
 Questions and Skepticism
 ------------------------
 
 #####*"Why did you write this?"*
-I used to use Unity, but the fickle state preservation of hotloading bothered me.
-I wanted a framework that let me use one language, with state-preserving runtime editing, with no overhead, with absurd amounts of static checking.
-
-I found Rust.
+I used to use Unity, but the complicated, fickle state preservation of their hotloading bothered me. I wanted a framework that was reliable, performant, easy to understand, and capable of being used for any kind of software project.
 
 #####*"So it's an abstracted main loop?"*
 An abstracted, hotloadable, state-preserving main loop!
 
 #####*"This isn't parallelizable."*
-Totally is. It was designed to be used with some form of [parallel job execution](https://github.com/mcpherrinm/parallel).
+Totally is. It was designed to be used with some form of parallel job execution, like [this](https://github.com/mcpherrinm/parallel).
 
 #####*"Isn't all data technically global?"*
 Schedules define what state a process can access, so unintended side effects would require exceptionally bad coding practices and probably copious amounts of alcohol.
 
 #####*"State data layout can't be modified at runtime!"*
-No, because that would invalidate the state. We *could* make it work though, see below.
+That would invalidate the state.
 
 #####*"Why GPL?"*
 Because I'm a nice guy. If you've got a license idea that's not MIT's anarchy or closed-source's despotism, hit me up!
 
-
-TODO / What can be improved?
-----------------------------
-
-#####Port to Windows and OSX
-I GUESS.
-
-#####Live reloading of state data.
-Closing the program to reset the state is awkward. It'd be nice if state::Data was hotloadable at runtime.
-The problem is, what subsystem owns the state::Data instance? What ungodly unsafe block could access functions of something it doesn't have type information for?
-
-Assuming the below isn't implemented, the cleanest way to achieve this would be to write an OS-level call to restart the application.
-
-#####Live _editing_ of state data's layout.
-Ownership of unknown types. Built-in serialization. Conversion function hooks on load. We go full erlang. PRs are welcome.
