+++ common/common.rs
@@ -1,6 +1,9 @@
 #[macro_use]
+extern crate lazy_static;
+#[macro_use]
 mod internal;
 
 pub mod hierarchy;
 pub mod system;
 pub mod settings;
+pub mod macros;
