+++ common/settings.rs
@@ -19,7 +19,3 @@ pub fn get_struct_lib_type() -> &'static str {
 pub fn get_state_lib_type() -> &'static str {
     "dylib"
 }
-
-pub fn get_macro_lib_type() -> &'static str {
-    "dylib"
-}
