--- macros/schedule/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-
-name = "schedule_macro"
-version = "0.0.1"
-authors = ["Kingsley Tolmie <kingsley.tolmie@gmail.com>"]
-
-[lib]
-name = "schedule_macro"
-path = "schedule_macro.rs"
-crate-type = ["dylib"]
