+++ state/state.rs
@@ -1,6 +1,6 @@
 // ATTENTION: This file is automatically generated. Don't modify it unless your life is terrible.
 #[macro_use]
-extern crate data_macro;
+extern crate common;
 
 // Now, lets get these imports out of the way...
 pub use simulation::SimulationState;
