+++ structs/graphics/Dependencies.toml
@@ -1,6 +1,3 @@
 # Note: local paths must be relative to /state/, not to this directory.
 
-[dependencies.data_macro]
-path = "./../macros/data"
-
 [dependencies.sdl2]
