+++ structs/simulation/Dependencies.toml
@@ -1,4 +1,1 @@
 # Note: local paths must be relative to /state/, not to this directory.
-
-[dependencies.data_macro]
-path = "./../macros/data"
