+++ tools/compile_scheduler/compile_scheduler.rs
@@ -44,11 +44,6 @@ fn main() {
     println!("Compiling scheduler");
 
     let mut command = io::Command::new(hierarchy::get_rustc_path().as_str().unwrap());
-    
-    // Link macro dirs
-    for path in hierarchy::get_all_macro_target_dirs().iter() {
-        command.arg("-L").arg(path.as_str().unwrap());
-    }
 
     // Link dependencies dirs
     for path in hierarchy::get_state_dependency_dirs().iter() {
