+++ tools/compile_state_library/compile_state_library.rs
@@ -36,8 +36,6 @@ fn main() {
     let target_path = current_dir.join("target");
     let source_filename = current_dir_name.to_string() + ".rs";
 
-    hierarchy::create_fresh_dir(&target_path).unwrap();
-
     println!("Compiling the State library");
 
     if should_update {
