+++ tools/generate_state_library/generate_cargo_toml.rs
@@ -5,8 +5,6 @@ use std::collections::HashMap;
 use toml;
 
 use common::hierarchy;
-use common::system;
-use common::settings;
 
 #[derive(RustcDecodable, RustcEncodable, Show)]
 struct TomlManifest {
@@ -116,9 +114,11 @@ pub fn exec(struct_src_dirs: &Vec<Path>) {
     // Fix the hashmap ordering
     for conflict in conflicts.iter_mut() {
         if conflict.manifest_a > conflict.manifest_b {
+        
             let tmp = conflict.manifest_a.clone();
             conflict.manifest_a = conflict.manifest_b.clone();
             conflict.manifest_b = tmp.clone();
+            
             let tmp = conflict.version_a.clone();
             conflict.version_a = conflict.version_b.clone();
             conflict.version_b = tmp.clone();
@@ -156,7 +156,7 @@ pub fn exec(struct_src_dirs: &Vec<Path>) {
     }
     
     let mut final_manifest = TomlManifest { dependencies: Some(HashMap::new()) };
-    for (manifest_name, manifest) in toml_manifests.iter() {
+    for (_, manifest) in toml_manifests.iter() {
         if manifest.dependencies.is_none() { 
             continue 
         }
@@ -165,8 +165,8 @@ pub fn exec(struct_src_dirs: &Vec<Path>) {
         }
     }
     
-    let cargo_toml_template_text =
-"#ATTENTION: This file is automatically generated. Don't modify it unless your life is terrible, or you wish it to be so.
+    let cargo_toml_text =
+format!("#ATTENTION: This file is automatically generated. Don't modify it unless your life is terrible, or you wish it to be so.
 [package]
 name = \"state\"
 version = \"0.0.1\"
@@ -177,20 +177,22 @@ name = \"state\"
 path = \"state.rs\"
 crate_type = [\"dylib\"]
 plugin = true
-";
+
+[dependencies.worldsong-common]
+path = \"{common_dir}\"
+
+{generated_text}
+", 
+common_dir = hierarchy::get_common_src_dir().as_str().unwrap(), 
+generated_text = toml::encode_str(&final_manifest));
+    
     
     let cargo_toml_path = hierarchy::get_state_src_dir().join("Cargo.toml");
+
+    println!("Creating new Cargo.toml");
     let mut cargo_toml_file = File::create(&cargo_toml_path);
-    cargo_toml_file.write_str(cargo_toml_template_text);
-    
-    let cargo_toml_generated_text = toml::encode_str(&final_manifest);
-    cargo_toml_file.write_str(cargo_toml_generated_text.as_slice());
-    cargo_toml_file.flush();
-    
-    // re-write the fields and values into a Cargo.toml
-    // Can I do this with an Encodable? :D
-    // Yep! Create a top-level TomlManifest and convert it to a string
-    // Append it to a default Cargo.toml file and we should be in business.
+    cargo_toml_file.write_str(cargo_toml_text.as_slice()).unwrap();
+    cargo_toml_file.flush().unwrap();
 }
 
 fn get_version(d: &TomlDependency) -> String {
