+++ tools/generate_state_library/generate_source.rs
@@ -26,11 +26,11 @@ pub fn exec(struct_src_dirs: &Vec<Path>) {
         type_names.push(struct_type_name)
     }
     
-    // Add the data_macro import
+    // Add the common lib import
     state_src_text.push_str(
 "// ATTENTION: This file is automatically generated. Don't modify it unless your life is terrible.
 #[macro_use]
-extern crate data_macro;\n\n"
+extern crate common;\n\n"
     );
     
     state_src_text.push_str("// Now, lets get these imports out of the way...\n");
@@ -65,8 +65,11 @@ extern crate data_macro;\n\n"
     state_src_text.push_str("    }\n");
     state_src_text.push_str("}\n");
     
+    let state_src_path = hierarchy::get_state_src_dir().join("state.rs");
+
     // save as state.rs
-    let mut state_src_file = File::create(&hierarchy::get_state_src_dir().join("state.rs"));
+    println!("Creating new state.rs");
+    let mut state_src_file = File::create(&state_src_path);
     state_src_file.write_str(state_src_text.as_slice()).unwrap();
     state_src_file.flush().unwrap();
 }
